rename: local-storage kebab-case로 이름 변경
